requirements.txt
discord.py
Flask
waitress
python-dotenv
aiosqlite
requests